add util funcs

+++ lib/net/ethernet.h
@@ -21,4 +21,4 @@ struct ethernet_footer
 };
 
 unsigned short ethernet_hlen(struct ethernet_header *header);
-void ethernet_initheader(struct ethernet_header *header, unsigned int type, unsigned char sha[ETHERNET_ADDRSIZE], unsigned char tha[ETHERNET_ADDRSIZE]);
+void ethernet_initheader(struct ethernet_header *header, unsigned short type, unsigned char sha[ETHERNET_ADDRSIZE], unsigned char tha[ETHERNET_ADDRSIZE]);
